making new room minigame
Made character double jump better. Wall jump might still have a tiny issue with it being consistent. working on tuck-n-roll mechanic.

Now there is a new room  that has 4 buttons to press. might be grounds for a minigame later on. Also in said room, i experimented with tiles. Tiles might not be something i might do a lot but could be of use for the meantime.