Tweaking around
Tweaked around in the animations of the buttons. The player will be able to press the buttons and be rerouted to the intended game mode. Once they are in the game, they can press esc to go back to the home screen. Made a couple objects folders to categorize the objects.