Big additions.
- Intro with theme
- music (janky)
- wall slip sprite

INTRO: made a room that runs the intro animation, when it ends it goes to title menu.

Music: all made b me, the music is made in audiotool. The music is yet to be implemented correctly.

Sprite: wall slip sprite took me a solid couple hours to figure out but it is okay for the meantime.