Title Screen Revamp (concept)
Used effects to make it as close to the idea i had in mind. The background moves using the cloud effect and some other effects. Tested them out till i realized 2 effects should be max per room to avoid lag.

Made a HOPBOY logo for the game, since i decided to call it HOPBOY. The name is subpar but anything to label the project for the meantime. It's cute and serves its purpose.

Added an exit button to the title screen so that the player doesnt have to press alt + f4 or navigate out of the game screen.

Music is still a major concern of mine since i have to make songs that loop correctly and figure out how to make the music in the game sound smooth and not jarring.

Next up: make a story intro animation and make the character navigate through a playable space that makes sense and not just some testing assets.

I also attempted to tinker around with the character's hitbox. It is very weird how the hit-box effects the character ohysics and momentum.